update RunAsync and consolidate implementation
