+++ libraries/Microsoft.Bot.Builder.Dialogs/DialogExtensions.cs
@@ -1,11 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Linq;
 using System.Security.Claims;
 using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Dialogs.Memory;
 using Microsoft.Bot.Builder.Skills;
 using Microsoft.Bot.Connector.Authentication;
 using Microsoft.Bot.Schema;
@@ -34,12 +36,61 @@ namespace Microsoft.Bot.Builder.Dialogs
 
             var dialogContext = await dialogSet.CreateContextAsync(turnContext, cancellationToken).ConfigureAwait(false);
 
+            // drop the return value - the whole point of this function is that the caller doesn't care
+            await InternalRunAsync(turnContext, dialog.Id, dialogContext, cancellationToken).ConfigureAwait(false);
+        }
+
+        internal static async Task<DialogTurnResult> InternalRunAsync(ITurnContext turnContext, string dialogId, DialogContext dialogContext, CancellationToken cancellationToken)
+        {
             // map TurnState into root dialog context.services
             foreach (var service in turnContext.TurnState)
             {
                 dialogContext.Services[service.Key] = service.Value;
             }
 
+            var dialogStateManager = new DialogStateManager(dialogContext);
+            await dialogStateManager.LoadAllScopesAsync(cancellationToken).ConfigureAwait(false);
+            dialogContext.Context.TurnState.Add(dialogStateManager);
+
+            DialogTurnResult dialogTurnResult = null;
+
+            // Loop as long as we are getting valid OnError handled we should continue executing the actions for the turn.
+            //
+            // NOTE: We loop around this block because each pass through we either complete the turn and break out of the loop
+            // or we have had an exception AND there was an OnError action which captured the error.  We need to continue the 
+            // turn based on the actions the OnError handler introduced.
+            var endOfTurn = false;
+            while (!endOfTurn)
+            {
+                try
+                {
+                    dialogTurnResult = await InnerRunAsync(turnContext, dialogId, dialogContext, cancellationToken).ConfigureAwait(false);
+
+                    // turn successfully completed, break the loop
+                    endOfTurn = true;
+                }
+                catch (Exception err)
+                {
+                    // fire error event, bubbling from the leaf.
+                    var handled = await dialogContext.EmitEventAsync(DialogEvents.Error, err, bubble: true, fromLeaf: true, cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                    if (!handled)
+                    {
+                        // error was NOT handled, throw the exception and end the turn. (This will trigger the Adapter.OnError handler and end the entire dialog stack)
+                        throw;
+                    }
+                }
+            }
+
+            // save all state scopes to their respective botState locations.
+            await dialogStateManager.SaveAllChangesAsync(cancellationToken).ConfigureAwait(false);
+
+            // return the redundant result because the DialogManager contract expects it
+            return dialogTurnResult;
+        }
+
+        private static async Task<DialogTurnResult> InnerRunAsync(ITurnContext turnContext, string dialogId, DialogContext dialogContext, CancellationToken cancellationToken)
+        {
             // Handle EoC and Reprompt event from a parent bot (can be root bot to skill or skill to skill)
             if (IsFromParentToSkill(turnContext))
             {
@@ -49,15 +100,13 @@ namespace Microsoft.Bot.Builder.Dialogs
                     if (!dialogContext.Stack.Any())
                     {
                         // No dialogs to cancel, just return.
-                        return;
+                        return new DialogTurnResult(DialogTurnStatus.Empty);
                     }
 
                     var activeDialogContext = GetActiveDialogContext(dialogContext);
 
                     // Send cancellation message to the top dialog in the stack to ensure all the parents are canceled in the right order. 
-                    await activeDialogContext.CancelAllDialogsAsync(true, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                    return;
+                    return await activeDialogContext.CancelAllDialogsAsync(true, cancellationToken: cancellationToken).ConfigureAwait(false);
                 }
 
                 // Handle a reprompt event sent from the parent.
@@ -66,11 +115,11 @@ namespace Microsoft.Bot.Builder.Dialogs
                     if (!dialogContext.Stack.Any())
                     {
                         // No dialogs to reprompt, just return.
-                        return;
+                        return new DialogTurnResult(DialogTurnStatus.Empty);
                     }
 
                     await dialogContext.RepromptDialogAsync(cancellationToken).ConfigureAwait(false);
-                    return;
+                    return new DialogTurnResult(DialogTurnStatus.Waiting);
                 }
             }
 
@@ -78,8 +127,10 @@ namespace Microsoft.Bot.Builder.Dialogs
             var result = await dialogContext.ContinueDialogAsync(cancellationToken).ConfigureAwait(false);
             if (result.Status == DialogTurnStatus.Empty)
             {
-                result = await dialogContext.BeginDialogAsync(dialog.Id, null, cancellationToken).ConfigureAwait(false);
+                result = await dialogContext.BeginDialogAsync(dialogId, null, cancellationToken).ConfigureAwait(false);
             }
+
+            await SendStateSnapshotTraceAsync(dialogContext, cancellationToken).ConfigureAwait(false);
 
             // Skills should send EoC when the dialog completes.
             if (result.Status == DialogTurnStatus.Complete || result.Status == DialogTurnStatus.Cancelled)
@@ -92,6 +143,23 @@ namespace Microsoft.Bot.Builder.Dialogs
                     await turnContext.SendActivityAsync(activity, cancellationToken).ConfigureAwait(false);
                 }
             }
+
+            return result;
+        }
+
+        /// <summary>
+        /// Helper to send a trace activity with a memory snapshot of the active dialog DC. 
+        /// </summary>
+        private static async Task SendStateSnapshotTraceAsync(DialogContext dialogContext, CancellationToken cancellationToken)
+        {
+            var traceLabel = dialogContext.Context.TurnState.Get<IIdentity>(BotAdapter.BotIdentityKey) is ClaimsIdentity claimIdentity && SkillValidation.IsSkillClaim(claimIdentity.Claims)
+                ? "Skill State"
+                : "Bot State";
+
+            // send trace of memory
+            var snapshot = GetActiveDialogContext(dialogContext).State.GetMemorySnapshot();
+            var traceActivity = (Activity)Activity.CreateTraceActivity("BotState", "https://www.botframework.com/schemas/botState", snapshot, traceLabel);
+            await dialogContext.Context.SendActivityAsync(traceActivity, cancellationToken).ConfigureAwait(false);
         }
 
         /// <summary>
