+++ libraries/Microsoft.Bot.Builder.Dialogs/DialogManager.cs
@@ -3,14 +3,9 @@
 
 using System;
 using System.Linq;
-using System.Security.Claims;
-using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Bot.Builder.Dialogs.Memory;
-using Microsoft.Bot.Builder.Skills;
-using Microsoft.Bot.Connector.Authentication;
-using Microsoft.Bot.Schema;
 using Newtonsoft.Json;
 
 namespace Microsoft.Bot.Builder.Dialogs
@@ -186,90 +181,13 @@ namespace Microsoft.Bot.Builder.Dialogs
             // Create DialogContext
             var dc = new DialogContext(Dialogs, context, dialogState);
 
-            // promote initial TurnState into dc.services for contextual services
-            foreach (var service in dc.Services)
-            {
-                dc.Services[service.Key] = service.Value;
-            }
-
-            // map TurnState into root dialog context.services
-            foreach (var service in context.TurnState)
-            {
-                dc.Services[service.Key] = service.Value;
-            }
-
-            // get the DialogStateManager configuration
-            var dialogStateManager = new DialogStateManager(dc, StateConfiguration);
-            await dialogStateManager.LoadAllScopesAsync(cancellationToken).ConfigureAwait(false);
-            dc.Context.TurnState.Add(dialogStateManager);
-
-            DialogTurnResult turnResult = null;
-
-            // Loop as long as we are getting valid OnError handled we should continue executing the actions for the turn.
-            //
-            // NOTE: We loop around this block because each pass through we either complete the turn and break out of the loop
-            // or we have had an exception AND there was an OnError action which captured the error.  We need to continue the 
-            // turn based on the actions the OnError handler introduced.
-            var endOfTurn = false;
-            while (!endOfTurn)
-            {
-                try
-                {
-                    if (context.TurnState.Get<IIdentity>(BotAdapter.BotIdentityKey) is ClaimsIdentity claimIdentity && SkillValidation.IsSkillClaim(claimIdentity.Claims))
-                    {
-                        // The bot is running as a skill.
-                        turnResult = await HandleSkillOnTurnAsync(dc, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        // The bot is running as root bot.
-                        turnResult = await HandleBotOnTurnAsync(dc, cancellationToken).ConfigureAwait(false);
-                    }
-
-                    // turn successfully completed, break the loop
-                    endOfTurn = true;
-                }
-                catch (Exception err)
-                {
-                    // fire error event, bubbling from the leaf.
-                    var handled = await dc.EmitEventAsync(DialogEvents.Error, err, bubble: true, fromLeaf: true, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                    if (!handled)
-                    {
-                        // error was NOT handled, throw the exception and end the turn. (This will trigger the Adapter.OnError handler and end the entire dialog stack)
-                        throw;
-                    }
-                }
-            }
-
-            // save all state scopes to their respective botState locations.
-            await dialogStateManager.SaveAllChangesAsync(cancellationToken).ConfigureAwait(false);
+            // Call the common dialog "continue/begin" execution pattern shared with the classic RunAsync extension method
+            var turnResult = await DialogExtensions.InternalRunAsync(context, _rootDialogId, dc, cancellationToken).ConfigureAwait(false);
 
             // save BotState changes
             await botStateSet.SaveAllChangesAsync(dc.Context, false, cancellationToken).ConfigureAwait(false);
 
             return new DialogManagerResult { TurnResult = turnResult };
-        }
-
-        /// <summary>
-        /// Helper to send a trace activity with a memory snapshot of the active dialog DC. 
-        /// </summary>
-        private static async Task SendStateSnapshotTraceAsync(DialogContext dc, string traceLabel, CancellationToken cancellationToken)
-        {
-            // send trace of memory
-            var snapshot = GetActiveDialogContext(dc).State.GetMemorySnapshot();
-            var traceActivity = (Activity)Activity.CreateTraceActivity("BotState", "https://www.botframework.com/schemas/botState", snapshot, traceLabel);
-            await dc.Context.SendActivityAsync(traceActivity, cancellationToken).ConfigureAwait(false);
-        }
-
-        private static bool IsFromParentToSkill(ITurnContext turnContext)
-        {
-            if (turnContext.TurnState.Get<SkillConversationReference>(SkillHandler.SkillConversationReferenceKey) != null)
-            {
-                return false;
-            }
-
-            return turnContext.TurnState.Get<IIdentity>(BotAdapter.BotIdentityKey) is ClaimsIdentity claimIdentity && SkillValidation.IsSkillClaim(claimIdentity.Claims);
         }
 
         // Recursively walk up the DC stack to find the active DC.
@@ -282,89 +200,6 @@ namespace Microsoft.Bot.Builder.Dialogs
             }
 
             return GetActiveDialogContext(child);
-        }
-
-        /// <summary>
-        /// Helper to determine if we should send an EndOfConversation to the parent or not.
-        /// </summary>
-        private static bool ShouldSendEndOfConversationToParent(ITurnContext context, DialogTurnResult turnResult)
-        {
-            if (!(turnResult.Status == DialogTurnStatus.Complete || turnResult.Status == DialogTurnStatus.Cancelled))
-            {
-                // The dialog is still going, don't return EoC.
-                return false;
-            }
-
-            if (context.TurnState.Get<IIdentity>(BotAdapter.BotIdentityKey) is ClaimsIdentity claimIdentity && SkillValidation.IsSkillClaim(claimIdentity.Claims))
-            {
-                // EoC Activities returned by skills are bounced back to the bot by SkillHandler.
-                // In those cases we will have a SkillConversationReference instance in state.
-                var skillConversationReference = context.TurnState.Get<SkillConversationReference>(SkillHandler.SkillConversationReferenceKey);
-                if (skillConversationReference != null)
-                {
-                    // If the skillConversationReference.OAuthScope is for one of the supported channels, we are at the root and we should not send an EoC.
-                    return skillConversationReference.OAuthScope != AuthenticationConstants.ToChannelFromBotOAuthScope && skillConversationReference.OAuthScope != GovernmentAuthenticationConstants.ToChannelFromBotOAuthScope;
-                }
-
-                return true;
-            }
-
-            return false;
-        }
-
-        private async Task<DialogTurnResult> HandleSkillOnTurnAsync(DialogContext dc, CancellationToken cancellationToken)
-        {
-            // the bot is running as a skill. 
-            var turnContext = dc.Context;
-
-            // Process remote cancellation
-            if (turnContext.Activity.Type == ActivityTypes.EndOfConversation && dc.ActiveDialog != null && IsFromParentToSkill(turnContext))
-            {
-                // Handle remote cancellation request from parent.
-                var activeDialogContext = GetActiveDialogContext(dc);
-
-                // Send cancellation message to the top dialog in the stack to ensure all the parents are canceled in the right order. 
-                return await activeDialogContext.CancelAllDialogsAsync(true, cancellationToken: cancellationToken).ConfigureAwait(false);
-            }
-
-            // Handle reprompt
-            // Process a reprompt event sent from the parent.
-            if (turnContext.Activity.Type == ActivityTypes.Event && turnContext.Activity.Name == DialogEvents.RepromptDialog)
-            {
-                if (dc.ActiveDialog == null)
-                {
-                    return new DialogTurnResult(DialogTurnStatus.Empty);
-                }
-
-                await dc.RepromptDialogAsync(cancellationToken).ConfigureAwait(false);
-                return new DialogTurnResult(DialogTurnStatus.Waiting);
-            }
-
-            // Continue execution
-            // - This will apply any queued up interruptions and execute the current/next step(s).
-            var turnResult = await dc.ContinueDialogAsync(cancellationToken).ConfigureAwait(false);
-            if (turnResult.Status == DialogTurnStatus.Empty)
-            {
-                // restart root dialog
-                turnResult = await dc.BeginDialogAsync(_rootDialogId, cancellationToken: cancellationToken).ConfigureAwait(false);
-            }
-
-            await SendStateSnapshotTraceAsync(dc, "Skill State", cancellationToken).ConfigureAwait(false);
-
-            if (ShouldSendEndOfConversationToParent(turnContext, turnResult))
-            {
-                // Send End of conversation at the end.
-                var code = turnResult.Status == DialogTurnStatus.Complete ? EndOfConversationCodes.CompletedSuccessfully : EndOfConversationCodes.UserCancelled;
-                var activity = new Activity(ActivityTypes.EndOfConversation)
-                {
-                    Value = turnResult.Result,
-                    Locale = turnContext.Activity.Locale,
-                    Code = code
-                };
-                await turnContext.SendActivityAsync(activity, cancellationToken).ConfigureAwait(false);
-            }
-
-            return turnResult;
         }
 
         /// <summary>
@@ -392,34 +227,6 @@ namespace Microsoft.Bot.Builder.Dialogs
                     }
                 }
             }
-        }
-
-        private async Task<DialogTurnResult> HandleBotOnTurnAsync(DialogContext dc, CancellationToken cancellationToken)
-        {
-            DialogTurnResult turnResult;
-
-            // the bot is running as a root bot. 
-            if (dc.ActiveDialog == null)
-            {
-                // start root dialog
-                turnResult = await dc.BeginDialogAsync(_rootDialogId, cancellationToken: cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                // Continue execution
-                // - This will apply any queued up interruptions and execute the current/next step(s).
-                turnResult = await dc.ContinueDialogAsync(cancellationToken).ConfigureAwait(false);
-
-                if (turnResult.Status == DialogTurnStatus.Empty)
-                {
-                    // restart root dialog
-                    turnResult = await dc.BeginDialogAsync(_rootDialogId, cancellationToken: cancellationToken).ConfigureAwait(false);
-                }
-            }
-
-            await SendStateSnapshotTraceAsync(dc, "Bot State", cancellationToken).ConfigureAwait(false);
-
-            return turnResult;
         }
     }
 }
